Added SongName Metadata, changeable by request input
Using form-data request, with keys file and songName and the user token in the header, this will append metadata to the file with the username of the uploader and the songName.

Solution to req.body.songName not being filled out in the GridFS storage section.
"if your file input element is in the top of form then every data below that gets assigned to req.body after your image is uploaded, in the result of which body remains empty while you upload photo.. So take file input element to very bottom of your form so that req.body will have data from other input elements"