Added API for Likes and Follows
Uses collections/like with a songID in the body, which is the document ID/filename, and takes the user token in the header to find the user to act on.
Uses collections/follow with artistName which is the username of the artist, and the token for the user in the header.